clean up and reorganize NotebookParserServer tests
