--- src/Microsoft.DotNet.Interactive.Documents.Tests/NotebookParserServerTests_Serialization.NotebookSerializeResponse_serialization_contract.approved.json
@@ -1,1 +1,0 @@
-{"rawData":"AQID","id":"the-id"}
